requirements.txt
Pillow
pystray
keyboard
pyzbar
pywin32; sys_platform == 'win32'
requests